requirements.txt
torch>=2.7.0
transformers>=4.40.0
opencv-python
pillow
numpy
pandas
scikit-image
scikit-learn
tqdm
matplotlib
huggingface_hub
